requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Pillow==10.0.0
python-dotenv==1.0.0
google-generativeai==0.3.1
opencv-python==4.8.0.74
pytesseract==0.3.10
numpy==1.24.3
matplotlib==3.7.2
pandas==2.0.3
Werkzeug==2.3.7
gunicorn==21.2.0

# Development dependencies (optional)
pytest==7.4.0
black==23.7.0
flake8==6.1.0
